ml_service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.2.2
torchvision==0.17.2
timm==0.9.7
Pillow>=9.0.1
numpy>=1.19.5
scikit-learn==1.3.0
scikit-image==0.21.0
requests==2.31.0
opencv-python==4.8.1.78
scipy==1.11.2
aiohttp==3.9.1
asyncio-throttle==1.0.2
ultralytics==8.0.232
fastapi==0.109.2
uvicorn==0.27.1
tensorflow-macos==2.16.1
python-multipart==0.0.5
pydantic==2.6.1
geopy==2.4.1 